Replace MongoDB course
Confirmed with MongoDB University that the edX course is retired.

Replaced it in the curriculum with two options, learners may
select one, the other, or both. One is more full stack oriented,
the other dives into backend a bit more. Both are worthwhile, either
is sufficient to prep for the FCC backend projects.

Updated CHANGELOG to reflect curriculum change.