Move MongoDB to the beginning of Tier 2 / clarify.
The MongoDB course is offered on a schedule and has assignments that unfold over the course of 7 weeks. Moved signing up for it to the beginning of Tier 2, with instructions to begin it during the next available session. Originally, this was an either/or course with Programming for the Web, but I've moved the latter into Tier 1 and am considering them a two-course sequence now as one focuses more on frontend, the other on backend. 

Closes #25