Added Software Construction: Object-Oriented Design
Added the next course in the How to Code series to the CS-WD and WD-CS guides. My intent is to include all courses in the series (they are fantastic) but this one just became available this week and the final one isn't available until 2018, so I'll add that one at that time.