Release version 4.0
Version 4.0
- Deprecates all prior guides.
- Migrates from coursera and edx
- Migrates towards opencourseware unlikely to hit paywalls
- provides a tiered 'core' main path
- provides opportunity to specialize in each tier
- Strongly focuses on core topics and 'reading the docs'
- Adds open source contributions
- Adds option to dive far more deeply into Algorithms

This revision is intended to accomplish two main goals:

1) Future-proof the guide by stepping away from  potential paywalls
2) Add clarity to avoid 'which guide should I choose'

A third goal involves restructuring with a core knowledge foundation.
This allows for diving deeply into a specialization
or sampling several specializations
all as part of a unified curriculum.

The focus remains on open source courses and reading materials.

The structure is changed to be slightly less linear
with courses/reading/practice each tier
but with the intention that the tier be a cohesive unit
with the learner taking courses, then reading and practicing
as fits their learning style.