Enhance the Odin Project entries
Removed the Reading resources for
the Odin Project specialization as they
are now behind a paywall.

I'll keep an eye out for replacements.
However, as a Specialization, the Odin
Project can stand on it's own as a
resource for now.

Split the Odin Project into two
separate specializations so as to
include the Full Stack JavaScript
option.

closes #159